+++ TrinketTinker/Effects/Abilities/NopAbility.cs
@@ -1,0 +1,18 @@
+using TrinketTinker.Effects.Proc;
+using TrinketTinker.Models;
+using TrinketTinker.Models.Mixin;
+
+namespace TrinketTinker.Effects.Abilities;
+
+/// <summary>Does absolutely nothing.</summary>
+public sealed class NopAbility(TrinketTinkerEffect effect, AbilityData data, int lvl)
+    : Ability<NoArgs>(effect, data, lvl)
+{
+    /// <summary>Does absolutely nothing.</summary>
+    /// <param name="proc"></param>
+    /// <returns></returns>
+    protected override bool ApplyEffect(ProcEventArgs proc)
+    {
+        return true;
+    }
+}